Move packages around so they make more sense in terms of OSGi exports.
