+++ libs/virtual-node/virtual-node-datamodel/src/integrationTest/kotlin/net/corda/libs/configuration/datamodel/tests/VNodeTestUtils.kt
@@ -3,8 +3,8 @@ package net.corda.libs.configuration.datamodel.tests
 import net.corda.db.core.DbPrivilege
 import net.corda.libs.configuration.datamodel.DbConnectionConfig
 import net.corda.libs.cpi.datamodel.CpiMetadataEntity
-import net.corda.libs.virtualnode.datamodel.HoldingIdentityEntity
-import net.corda.libs.virtualnode.datamodel.VirtualNodeEntity
+import net.corda.libs.virtualnode.datamodel.entities.HoldingIdentityEntity
+import net.corda.libs.virtualnode.datamodel.entities.VirtualNodeEntity
 import net.corda.orm.utils.transaction
 import net.corda.test.util.TestRandom
 import net.corda.v5.base.types.MemberX500Name
