+++ libs/virtual-node/virtual-node-datamodel/src/main/kotlin/net/corda/libs/virtualnode/datamodel/VirtualNodeEntities.kt
@@ -1,5 +1,8 @@
 package net.corda.libs.virtualnode.datamodel
 
+import net.corda.libs.virtualnode.datamodel.entities.HoldingIdentityEntity
+import net.corda.libs.virtualnode.datamodel.entities.VirtualNodeEntity
+
 object VirtualNodeEntities {
     val classes = setOf(
         HoldingIdentityEntity::class.java,
