+++ libs/virtual-node/virtual-node-datamodel/src/main/kotlin/net/corda/libs/virtualnode/datamodel/entities/HoldingIdentityEntity.kt
@@ -1,4 +1,4 @@
-package net.corda.libs.virtualnode.datamodel
+package net.corda.libs.virtualnode.datamodel.entities
 
 import net.corda.db.schema.DbSchema.CONFIG
 import net.corda.db.schema.DbSchema.HOLDING_IDENTITY_DB_TABLE
